requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==2.8
Django==4.1.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
idna==3.4
numpy==1.23.2
psycopg2==2.9.5
pycparser==2.21
pyOpenSSL==23.0.0
pyproj==3.4.0
PySocks==1.7.1
python-polylabel==0.6
pytz==2022.2.1
requests==2.22.0
Shapely==1.8.5.post1
six==1.14.0
sqlparse==0.4.3
typing_extensions==4.3.0
tzlocal==4.2
urllib3==1.26.12


